requirements.txt
numpy~=2.2.5
flask~=3.1.1
pillow~=11.2.1
scikit-learn~=1.6.1